Refactor cp/mv shared code
